fix(filemanager) change i32 to i64 to support bigger sizes (#218)
* fix(filemanager): change i32 to i64 to support bigger sizes

* test(filemanager): test max i64 deserialization

+++ lib/workload/stateless/filemanager/database/queries/ingester/aws/update_reordered_for_created.sql
@@ -10,7 +10,7 @@ with input as (
         $2::text[],
         $3::text[],
         $4::timestamptz[],
-        $5::integer[],
+        $5::bigint[],
         $6::text[],
         $7::timestamptz[],
         $8::text[],
